requirements.txt
# Django Dependancy
django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9

# Healingmentor Backend Dependancy
gunicorn==21.2.0
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.0